+++ eq_employee_documents/RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 ## Modul eq_employee_documents
 
+#### 07.08.2018
+#### Version 1.0.1
+##### ADD
+- Created folder for file format icons
+- Overriding create function in hr.employee.document model for storing the images in documents
+- Showing icon image in kanban view
+- Using widget = 'image' from oca module (web_tree_image) to show images in tree view
+
 #### 06.08.2018
 #### Version 1.0.0
 ##### ADD